job limit is usize

+++ src/worker.rs
@@ -27,7 +27,7 @@ pub trait Worker {
   fn source(&self) -> String;
   fn sink(&self) -> String;
 
-  fn start(&self, limit : Option<i32>) -> Result<(), Error> {
+  fn start(&self, limit : Option<usize>) -> Result<(), Error> {
     let mut work_counter = 0;
     // Connect to a task ventilator
     let mut context_source = Context::new();
